修改活动id数据类型，删除冗余字段。

+++ driving-api/src/main/java/com/udriving/drivingapi/controller/request/UploadFlockQrCodeRequestParameter.java
@@ -10,7 +10,7 @@ public class UploadFlockQrCodeRequestParameter {
     /**
      * 活动id
      */
-    private Integer acitivityId;
+    private Long acitivityId;
 
     /**
      * 二维码文件的Base64编码字符串
